Breaking out classes
