+++ src/Rn.NetCore.Metrics/Builders/ICoreMetricBuilder.cs
@@ -1,0 +1,11 @@
+﻿using System;
+using Rn.NetCore.Metrics.Models;
+
+namespace Rn.NetCore.Metrics.Builders;
+
+public interface ICoreMetricBuilder<TBuilder>
+{
+  ICoreMetricBuilder<TBuilder> AddAction(Action<CoreMetric> action);
+  void SetSuccess(bool success);
+  CoreMetric Build();
+}
